Python/Django REST Framework/simple_buy/project/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
djoser==2.2.0
psycopg2==2.9.7
drf-nested-routers==0.93.4
python-dotenv==1.0.0